First commit.
Added a script I've been using to just ease the copy/pasting of
hex bytes only - stripping out the extra characters I dont care about.

Also added a modified template for loading ZIP files, essentially I wanted
to be able to load APK files - which are the same structually as JAR files.
This is just a quick hack on top of the ZipTemplateAdv from the SweetScapes
site, just to aid in some of the weirdnesses that are introduced when parsing
JARs.