Fixing the extra-small padding issue with extra fields.
There was a specific ZIP that had some extra odd padding,
essentially it was less than 2 bytes long, which cause issues
when attempting to create an EXTRAFIELD struct. Now the struct
will just not be created and a padding object will be. Other
usage is fixed to not reference the struct if it was never
created.