MPQTemplate added.
Started some work on an MPQ template -- not fully finished, but
it filled the small need that I had for it. Probably won't put
much time into it after this, just submitting so I don't lose it
and if anyone else ever needs it they might be able to find it.