MachOTemplate added.
Added an early version of the macho template. This works for the
purposes that I needed it for, mainly on macho32 bit binaries. It
has a bit of work that needs to be done, will slowly add these things
and fix bugs as I come across them. Figured it'd be better to submit
this now before I forget all about it - maybe someone will have a use
for it :)