Was fooling around with a Syntax Highlighter for Smali
This is in no ways complete or even that great - the emacs schema is
much more complete and functional. Was just toying around with the syntax
highlighting engine for 010Editor in the newest beta. It has some bugs I
think, but I won't be using this for smali editing/browsing anyway. Oddly
enough you can't explort only ONE syntax for highlighting/importing. So
you'll need to "Import List..." and then choose to import the Smali one
only. Though the others are the defaults so it should be safe unless you
edited the others to just blow them away. If you've edited the others
please let me know so I can figure out wtf is actually going on.