Fixed how warnings are displayed.
A "Warning" in 010Editor puts text into the status bar,
not the "output" box. If more than one warning occurs,
you lose the data which was in the status bar. So I've
done a quick hack to output warnings to the output box
and summerize the number of warnings in the status bar.

This change will also list the template execution as
a failure, with the number of warnings occured - opposed
to "successfully executed".